Moving to lib/ directory
